added hostname website location, and cleaned up some console.log calls
I think that with promises made in the root app, on promise fullfillment
I've read somewhere that react re-renders. The only way around this is
actually using the prop in the child components when the promise I
assume is fulfilled. Using any of the other state functions in children to
fetch these props, such as componentWillMount, componentDidMount, or
componentWillReceiveProps doesn't update props being passed. They will
only correctly show up on render if they are called in the render location.

This makes for a not so nice looking render call, but it works. I'll look
into this later on and if I can find any indication that there is a better
ay, I'll update. These props that are to be loaded in the child components
are of course, state variables in their parent component that are in turn
created by a promise in the parent component.